refactor: split top-5 monolithic route files into modular directories
Following the established routes/<name>/{...,index}.ts pattern used by
auth/, bins/, and export/, split:

- aiStream.ts (583 lines, 11 endpoints) into aiStream/{text,image,
  reorganize,helpers,index}
- adminSystem.ts (607 lines, 14 endpoints) into adminSystem/{maintenance,
  announcements,health,auditLog,locations,bins,index}
- admin.ts (616 lines, 12 endpoints) into admin/{users,userActions,
  registration,backup,index}
- tags.ts (453 lines) into tags/{list,single,bulkApply,bulkOps,helpers,
  index}
- items.ts (447 lines) into items/{list,bulk,index}

Hoisted getRegistrationMode() out of the routes layer into a dedicated
lib/registrationMode.ts (3 callers: auth/status, auth/register,
lib/oauth) so routes no longer reach into a sibling route file.

Tests that inspect router.stack updated to import sub-routers directly
where their target route lives.

Reuse pass: replaced 3 inlined "fetch user, throw if not found"
patterns with the existing assertUserExists() helper from
lib/adminHelpers.ts; deduplicated the multer fields config in
aiStream/image.ts via the shared analyzeImageFields constant.
