+++ server/src/index.ts
@@ -25,11 +25,11 @@ import { maintenanceGate } from './middleware/maintenance.js';
 import { requestLogger } from './middleware/requestLogger.js';
 import { requireActiveSubscription } from './middleware/requirePlan.js';
 import activityRoutes from './routes/activity.js';
-import { adminRoutes } from './routes/admin.js';
+import adminRoutes from './routes/admin/index.js';
 import { adminSecurityRoutes } from './routes/adminSecurity.js';
-import { adminSystemRoutes } from './routes/adminSystem.js';
+import adminSystemRoutes from './routes/adminSystem/index.js';
 import aiRoutes from './routes/ai.js';
-import { streamRouter as aiStreamRoutes } from './routes/aiStream.js';
+import aiStreamRoutes from './routes/aiStream/index.js';
 import apiKeysRoutes from './routes/apiKeys.js';
 import areasRoutes from './routes/areas.js';
 import attachmentsRoutes from './routes/attachments.js';
@@ -45,7 +45,7 @@ import customFieldsRoutes from './routes/customFields.js';
 import exportRoutes from './routes/export/index.js';
 import { internalRoutes } from './routes/internal.js';
 import itemCheckoutsRoutes, { locationCheckoutsRouter as locationCheckoutsRoutes } from './routes/itemCheckouts.js';
-import itemsRoutes from './routes/items.js';
+import itemsRoutes from './routes/items/index.js';
 import locationsRoutes from './routes/locations.js';
 import photosRoutes from './routes/photos.js';
 import { planRoutes } from './routes/plan.js';
@@ -58,7 +58,7 @@ import binShoppingListRoutes, {
   shoppingListRouter as shoppingListRoutes,
 } from './routes/shoppingList.js';
 import tagColorsRoutes from './routes/tagColors.js';
-import tagsRoutes from './routes/tags.js';
+import tagsRoutes from './routes/tags/index.js';
 import userPreferencesRoutes from './routes/userPreferences.js';
 
 const STATIC_DIR = path.join(import.meta.dirname, '..', 'public');
